feat(core, kernels): move utility functions to a new dedicated module
