+++ honeycomb-kernels/src/lib.rs
@@ -16,3 +16,4 @@ pub mod grisubal;
 pub mod remeshing;
 pub mod skewness;
 pub mod triangulation;
+pub mod utils;
