+++ honeycomb-kernels/src/utils/mod.rs
@@ -1,0 +1,9 @@
+//! collection of utilities for meshing algorithms
+
+mod anchors;
+mod routines;
+
+pub use anchors::{
+    BodyIdType, CurveIdType, EdgeAnchor, FaceAnchor, NodeIdType, SurfaceIdType, VertexAnchor,
+};
+pub use routines::is_orbit_orientation_consistent;
